project provided through Code Academy
BUILD PYTHON WEB APPS WITH FLASK
Receipts for Lovely Loveseats
(Question Story) We’ve decided to pursue the dream of small-business ownership and open up a furniture store called Lovely Loveseats for Neat Suites on Fleet Street. With your new found knowledge of Python programming, you are going to build a system to help speed up the process of creating receipts for your customers. In this project, we will be storing the names and prices of a furniture store’s catalog in variables. You will then process the total price and item list of customers, printing them to the output terminal.